// src/services/types/data.ts
export interface IIngredient {
    _id: string;
    name: string;
    type: string;
    proteins: number;
    fat: number;
    carbohydrates: number;
    calories: number;
    price: number;
    image: string;
    image_mobile: string;
    image_large: string;
    __v: number;
    constructorItemId?: string;
  }
// ...
  export interface IAddIngredient {
    readonly type: typeof ADD_INGREDIENT;
    readonly constructorItemId: string;
    readonly payload: IIngredient;
  }

  export interface IDeleteIngredient {
    readonly type: typeof DELETE_INGREDIENT;
    readonly payload: IIngredient;
  }

  export interface IConstructorInitialState {
    constructorBunElement: IIngredient | undefined;
    constructorFillingList: Array<IIngredient>;
  }

  export interface ISortIngredients {
    readonly type: typeof SORT_INGREDIENTS;
    payload: Array<IIngredient>;
  }

  export type TConstructorAction =
  | IAddIngredient
  | ISortIngredients
  | IDeleteIngredient
  | IResetIngredient;
// ...
  export interface IResetIngredient {
    readonly type: typeof RESET_INGREDIENT;
  }